+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Implement `HTMLTemplateElement` and parse `<template>` tags.
 * Properly parse the `<main>` tag.
 * Remove support for the non-standard `<command>` tag.
+* Create `HTMLCanvasElement` when parsing `<canvas>` tags.
 
 # domino 1.0.23 (30 Jan 2016)
 * Fix `CSSStyleDeclaration#setProperty`. (#71)